requirements.txt
typer==0.7.0
rich==13.3.2
black==23.1.0
isort==5.12.0
openai==0.27.1
python-dotenv==1.0.0
pytest==7.2.2
